requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
facexlib==0.3.0
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
gfpgan==1.3.8
gitdb==4.0.11
gitpython==3.1.43
gputil==1.4.0
grpcio==1.68.0
idna==3.10
imageio==2.36.0
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
lazy-loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
loguru==0.7.2
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
numba==0.60.0
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
-e file:///root/pixelupbench
platformdirs==4.3.6
prompt-toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.1
pydantic-core==2.27.1
pydantic-settings==2.6.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
ruff==0.8.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tb-nightly==2.19.0a20241121
tensorboard-data-server==0.7.2
tifffile==2024.9.20
tinycss2==1.4.0
torch==2.1.0
torchvision==0.16.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.1.0
typing-extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.1.3
yapf==0.43.0
yarg==0.1.9
